requirements.txt
Django==4.2
djangorestframework==3.14.0
gunicorn
celery==5.3.1
redis==5.0.1
Pillow
djangorestframework-simplejwt
